Petroleum Oil Spills Impacting Navigable US Waters

A dataset containing yearly gallons spilled, number of incidents, and source type of oil spills from 1985, 1990, 1995-2020.

Usage

oilSpills

Format

A data frame with 280 rows and 4 variables:

Year: year of spill

Source: source of spill

Incidents: number of incidents of the source for that year

Gallons spilled: number of gallons of oil spilled of that source for that year

Source

https://www.bts.gov/content/petroleum-oil-spills-impacting-navigable-us-waters

Examples

oilSpills |> dplyr::group_by(Year) |> dplyr::summarize(Gallons = sum(`Gallons spilled`))
#> # A tibble: 28 × 2
#>     Year  Gallons
#>    <dbl>    <dbl>
#>  1  1985 24985636
#>  2  1990 23625644
#>  3  1995  7858833
#>  4  1996  9324985
#>  5  1997  2767292
#>  6  1998  2638557
#>  7  1999  3472754
#>  8  2000  4270144
#>  9  2001  2549419
#> 10  2002  1726019
#> # … with 18 more rows